fix install script

+++ install-scripts/sddm.sh
@@ -2,6 +2,7 @@
 # Install and configure SDDM Login Manager
 
 # login managers to attempt to disable
+LOG="Install-Logs/install-$(date +%d-%H%M%S)-sddm.log"
 login=(
   lightdm
   gdm3
@@ -15,14 +16,12 @@ SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 # Change the working directory to the parent directory of the script
 PARENT_DIR="$SCRIPT_DIR/.."
-cd "$PARENT_DIR" || { echo "${ERROR} Failed to change directory to $PARENT_DIR"; exit 1; }
+cd "$PARENT_DIR" || { echo "[ERROR] Failed to change directory to $PARENT_DIR" | tee -a "$LOG"; exit 1; }
 
-# Set the name of the log file to include the current date and time
-LOG="Install-Logs/install-$(date +%d-%H%M%S)_sddm.log"
 
 # Source the global functions script
 if ! source "$SCRIPT_DIR/base.sh"; then
-  echo "${ERROR} Failed to source ${ORANGE}base.sh\n${RESET}" | tee -a "$LOG"
+  echo "[ERROR] Failed to source base.sh\n" | tee -a "$LOG"
   exit 1
 fi
 
@@ -50,17 +49,23 @@ printf "${INFO} Installing ${SKY_BLUE}Additional SDDM Theme${RESET}\n" | tee -a 
 sudo mkdir -p /usr/share/sddm/themes || { printf "%s - Failed to create sddm themes dir\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
 sudo cp -an $PARENT_DIR/assets/sddm/. /usr/share/sddm/themes || { printf "%s - Failed to copy ${YELLOW}Themes${RESET} to sddm themes dir\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
 sudo mkdir -p /etc/sddm.conf.d || { printf "%s - Failed to create sddm config dir\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
-sudo cp -an /etc/sddm.conf.d/sddm.conf /etc/sddm.conf.d/sddm.conf.bak || { printf "%s - Failed to backup ${YELLOW}sddm config${RESET}\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
-sudo cp -an /usr/lib/sddm/sddm.conf.d/default.conf /etc/sddm.conf.d/sddm.conf || { printf "%s - Failed to copy ${YELLOW}default sddm config${RESET}\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
+
+if [ -f "/etc/sddm.conf.d/sddm.conf" ]; then
+  sudo cp -an /etc/sddm.conf.d/sddm.conf /etc/sddm.conf.d/sddm.conf.bak || { printf "%s - Failed to backup ${YELLOW}sddm config${RESET}\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
+else
+  sudo cp -an /usr/lib/sddm/sddm.conf.d/default.conf /etc/sddm.conf.d/sddm.conf || { printf "%s - Failed to copy ${YELLOW}default sddm config${RESET}\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
+fi
+
 sudo sed -i "s|.*Current=.*|Current=eucalyptus-drop|" /etc/sddm.conf.d/sddm.conf  || { printf "%s - Failed to edit ${YELLOW}sddm config${RESET} to include Themes\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
-sudo ln -sf /etc/sddm.conf.d/sddm.conf /etc/sddm.conf || || { printf "%s - Failed to link ${YELLOW}sddm config${RESET}for support\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
+sudo ln -sf /etc/sddm.conf.d/sddm.conf /etc/sddm.conf || { printf "%s - Failed to link ${YELLOW}sddm config${RESET}for support\n" "${ERROR}" | tee -a "$LOG"; exit 1; }
 
 printf "\n%.0s" {1..1} | tee -a "$LOG"
 
 printf "${INFO} Activating sddm service........\n" | tee -a "$LOG"
-sudo systemctl enable sddm
+sudo systemctl enable --now sddm
 
 wayland_sessions_dir=/usr/share/wayland-sessions
 [ ! -d "$wayland_sessions_dir" ] && { printf "$CAT - $wayland_sessions_dir not found, creating...\n"; sudo mkdir "$wayland_sessions_dir" 2>&1 | tee -a "$LOG"; }
 
-printf "\n%.0s" {1..2}
+printf "${OK} Finished configuring ${SKYBLUE}SDDM${RESET}\n" | tee -a "$LOG"
+printf "\n%.0s" {1..1}
